requirements.txt
addicted3==3.0.0
androguard==3.3.5
appdirs==1.4.4
appnope==0.1.2
asn1crypto==1.4.0
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
backports-abc==0.5
build==0.3.1.post1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cycler==0.10.0
dbr==8.1.2
decorator==4.4.2
distlib==0.3.1
elementpath==2.2.0
filelock==3.0.12
future==0.18.2
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.3.1
libusb1==1.9.2
lxml==4.6.3
manafa>=0.3.5
matplotlib==3.3.4
monsoon==0.1.88
networkx==2.5
noattr==0.0.9
numpy==1.19.4
packaging==20.9
pandas==1.1.5
parso==0.8.1
patsy==0.5.1
pep517==0.10.0
pexpect==4.8.0
physalia==0.0.1.dev122
pickleshare==0.7.5
Pillow==8.2.0
pip-autoremove==0.9.1
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
Pygments==2.8.0
PyHamcrest==2.0.2
pylatexenc==2.9
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
python-textops3==3.1.0
pytz==2021.1
pyusb==1.1.1
requests==2.25.1
scipy==1.6.1
six==1.15.0
statsmodels==0.12.2
tabulate==0.8.9
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
txaio==20.12.1
tzlocal==2.1
urllib3==1.26.5
virtualenv==20.4.2
wcwidth==0.2.5
whichcraft==0.6.1
zope.interface==5.2.0